3.1.7 - Renamed list package to stack
